# frontend/openapi/homecenter-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Home Media Center — API",
    "description": "REST interface for the Android TV client. Media is split into three\ncategories (VIDEO, PHOTO, AUDIO). The server reads from its local index\nand forwards files from Samba through /api/v1/media/{id}/stream with\nRange request support.\n\nEverything except /api/v1/auth/login requires the\n`Authorization: Bearer <token>`. A token is issued by logging in\nwith a username and password, or a username and PIN.\n",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "v1"
  },
  "security": [
    {
      "bearer": []
    }
  ],
  "tags": [
    {
      "name": "Scanning",
      "description": "Starting a Samba scan and its history"
    },
    {
      "name": "Library",
      "description": "Overview and listing of indexed media"
    },
    {
      "name": "Playback",
      "description": "Transfers files from Samba with Range request support"
    },
    {
      "name": "Authentication",
      "description": "Issuing and revoking tokens for the TV client"
    }
  ],
  "paths": {
    "/api/v1/scan": {
      "get": {
        "tags": [
          "Scanning"
        ],
        "summary": "Scan history across all sources",
        "operationId": "history",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScanRunDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Scanning"
        ],
        "summary": "Starts a scan in the background and returns at once; without sourceId it walks every enabled source",
        "operationId": "start",
        "parameters": [
          {
            "name": "sourceId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ScanStartedDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Revokes the token this request is signed with",
        "operationId": "logout",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Logs in with a username and password, or a username and PIN, and returns a token",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scan/latest": {
      "get": {
        "tags": [
          "Scanning"
        ],
        "summary": "The latest scan, including the progress of one still running",
        "operationId": "latest",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ScanRunDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/media": {
      "get": {
        "tags": [
          "Library"
        ],
        "summary": "Lists media, optionally filtered by category, source, genre and search text",
        "operationId": "list",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "VIDEO",
                "PHOTO",
                "AUDIO"
              ]
            }
          },
          {
            "name": "sourceId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "genreId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 60
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MediaPageDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/media/{id}": {
      "get": {
        "tags": [
          "Library"
        ],
        "summary": "Details of a single item",
        "operationId": "detail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MediaItemDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/media/{id}/stream": {
      "get": {
        "tags": [
          "Playback"
        ],
        "summary": "Plays a file; with a Range header it returns 206 and the requested part",
        "operationId": "stream",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Range",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/media/{id}/poster": {
      "get": {
        "tags": [
          "Library"
        ],
        "summary": "Locally cached poster or episode still",
        "operationId": "poster",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/library": {
      "get": {
        "tags": [
          "Library"
        ],
        "summary": "Library overview—the counts behind the Videos / Photos / Music tiles",
        "operationId": "library",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LibrarySummaryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/genres": {
      "get": {
        "tags": [
          "Library"
        ],
        "summary": "Movie genres currently in use in the library",
        "operationId": "genres",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaGenreDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Who is logged in with this token",
        "operationId": "me",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ScanStartedDto": {
        "type": "object",
        "properties": {
          "queuedSources": {
            "type": "integer",
            "format": "int32",
            "description": "How many sources will be walked"
          },
          "sources": {
            "type": "array",
            "description": "Source names in the order they will be walked",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "LoginRequestDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1
          },
          "secret": {
            "type": "string",
            "description": "Password, or PIN if the user has one set",
            "minLength": 1
          },
          "deviceName": {
            "type": "string"
          }
        },
        "required": [
          "secret",
          "username"
        ]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "USER"
            ]
          },
          "hasPin": {
            "type": "boolean"
          }
        }
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Shown by the server exactly once; it cannot be displayed again"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          }
        }
      },
      "ScanRunDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "trigger": {
            "type": "string",
            "enum": [
              "MANUAL",
              "SCHEDULED",
              "STARTUP"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "RUNNING",
              "COMPLETED",
              "FAILED"
            ]
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "durationSeconds": {
            "type": "integer",
            "format": "int64"
          },
          "directoriesScanned": {
            "type": "integer",
            "format": "int32"
          },
          "filesSeen": {
            "type": "integer",
            "format": "int32"
          },
          "itemsAdded": {
            "type": "integer",
            "format": "int32"
          },
          "itemsUpdated": {
            "type": "integer",
            "format": "int32"
          },
          "itemsRemoved": {
            "type": "integer",
            "format": "int32"
          },
          "errorMessage": {
            "type": "string"
          }
        }
      },
      "MediaGenreDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "MediaItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "sourceId": {
            "type": "integer",
            "format": "int64"
          },
          "category": {
            "type": "string",
            "enum": [
              "VIDEO",
              "PHOTO",
              "AUDIO"
            ]
          },
          "title": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "relativePath": {
            "type": "string"
          },
          "extension": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time"
          },
          "contentType": {
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/MediaMetadataDto"
          },
          "streamUrl": {
            "type": "string"
          }
        }
      },
      "MediaMetadataDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "MOVIE",
              "TV_EPISODE"
            ]
          },
          "description": {
            "type": "string"
          },
          "releaseYear": {
            "type": "integer",
            "format": "int32"
          },
          "rating": {
            "type": "number",
            "format": "double"
          },
          "groupKey": {
            "type": "string"
          },
          "groupTitle": {
            "type": "string"
          },
          "seasonNumber": {
            "type": "integer",
            "format": "int32"
          },
          "episodeNumber": {
            "type": "integer",
            "format": "int32"
          },
          "partNumber": {
            "type": "integer",
            "format": "int32"
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaGenreDto"
            }
          },
          "posterUrl": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "providerId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "MediaPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaItemDto"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "limit": {
            "type": "integer",
            "format": "int32"
          },
          "offset": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CategoryTileDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "enum": [
              "VIDEO",
              "PHOTO",
              "AUDIO"
            ]
          },
          "label": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "LibrarySummaryDto": {
        "type": "object",
        "properties": {
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryTileDto"
            }
          },
          "totalItems": {
            "type": "integer",
            "format": "int64"
          },
          "totalSizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "lastScan": {
            "$ref": "#/components/schemas/ScanRunDto"
          }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "description": "Token from POST /api/v1/auth/login",
        "scheme": "bearer"
      }
    }
  }
}